feat(WriteDiaryContentActivity.kt): 힌트 버튼의 동작을 데이터 바인딩으로 바꿈
